+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # CHANGELOG
 
 ## 0.8.2-SNAPSHOT
+* Add PropertyUsageOption.REQUIRED to describe fields that are required.
+
+* Add PropertyUsageOption.UNIQUE_IDENTIFIER to describe fields that are
+  unique identifiers.
+
+* Add validateState(description, state) to Utils to validate fields that
+  are required. If the field is null, REQUIRED and UNIQUE_IDENTIFIER, a UUID is
+  automatically generated. If the field is null and REQUIRED, an Exception is thrown.
 
 * Add PropertyUsageOption.SENSITIVE to describe fields that contain sensitive
   information. When marked, the field will be hidden when serializing to JSON